Refactor unit tests into a dedicated file
